% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adapt_rho.R
\name{Adapt_rho}
\alias{Adapt_rho}
\title{Calculate adaptive step-size for Alternating Direction Method of Multipliers (ADMM) algorithm}
\usage{
Adapt_rho(rho, r_norm, s_norm, tau = 2, eta = 10)
}
\arguments{
\item{rho}{\link{numeric}: (Fixed) ADMM step size}

\item{r_norm}{\link{numeric}: L2-norm of primal residuals}

\item{s_norm}{\link{numeric}: L2-norm of dual residuals}

\item{tau}{}

\item{eta}{}
}
\value{
rho \link{numeric}: Adaptive ADMM step size
}
\description{
R-function implementing adaptive step-size for ADMM
}
