% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_ll_fisher.R
\name{cox_ll_fisher}
\alias{cox_ll_fisher}
\title{cox_ll_fisher}
\usage{
cox_ll_fisher(X, d, beta, Riskset)
}
\arguments{
\item{X}{\link{matrix}: Regression matrix of dimension n_obs x p_vars}

\item{d}{\link{data frame}: Data set with variables Tstart, Tstop, trans and status}

\item{beta}{\link{vector}: Regression parameter}

\item{Riskset}{\link{list}: Risk set list}
}
\value{
info: Fisher information matrix at beta
}
\description{
R-function calculating Fisher information matrix of partial log-likelihood
}
