% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_ll_lagr_gradient.R
\name{cox_ll_lagr_gradient}
\alias{cox_ll_lagr_gradient}
\title{cox_ll_lagr_gradient}
\usage{
cox_ll_lagr_gradient(X, d, beta, Riskset, rho = 1, theta, nu)
}
\arguments{
\item{X}{\link{matrix}: Design matrix of dimension n_long x p_vars}

\item{d}{\link{data frame}: Data set with variables Tstart, Tstop, trans and status}

\item{beta}{\link{vector}: Regression parameter}

\item{Riskset}{\link{list}: Risk set list}

\item{rho}{\link{numeric}: Augmented Lagrangian parameter (ADMM step size; default: 1)}

\item{theta}{\link{vector}: ADMM parameter theta (dimension M x 1)}

\item{nu}{\link{vector}: ADMM parameter nu (dimension M x 1)}
}
\value{
scorevector: Gradient at beta
}
\description{
R-function calculating the gradient of the augmented Lagrangian form of full/partial log-likelihood
}
