% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_ll_lagr_gradient_K.R
\name{cox_ll_lagr_gradient_K}
\alias{cox_ll_lagr_gradient_K}
\title{Calculate gradient of the augmented Lagrangian form of log-likelihood function}
\usage{
cox_ll_lagr_gradient_K(X, d, K, beta, Riskset, rho = 1, theta, nu)
}
\arguments{
\item{X}{\link{matrix}: Design matrix of dimension n_long x p_vars}

\item{d}{\link{data frame}: Data set with variables Tstart, Tstop, trans and status}

\item{K}{\link{matrix}: Penalty matrix of dimension M x p}

\item{beta}{\link{vector}: Regression parameter}

\item{Riskset}{\link{list}: Risk set list}

\item{rho}{\link{numeric}: Augmented Lagrangian parameter (ADMM step size; default: 1)}

\item{theta}{\link{vector}: ADMM parameter theta (dimension M x 1)}

\item{nu}{\link{vector}: ADMM parameter nu (dimension M x 1)}
}
\value{
scorevector: Score function at beta
}
\description{
R-function calculating the gradient of the augmented Lagrangian form of full/partial log-likelihood with penalty matrix K
}
