% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_ll.R
\name{full_ll}
\alias{full_ll}
\title{full_ll}
\usage{
full_ll(X, d, beta, Riskset)
}
\arguments{
\item{X}{\link{matrix}: Regression matrix of dimension n_obs x p_vars}

\item{d}{\link{data frame}: Data set with variables Tstart, Tstop, trans and status}

\item{beta}{\link{vector}: Regression parameter}

\item{Riskset}{\link{matrix}: Risk set matrix}
}
\value{
loglik \link{numeric}: Negative full log-likelihood at beta
}
\description{
R-function calculating the negative full log-likelihood function
}
