% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_contrast_matrix.R
\name{pairwise_contrast_matrix}
\alias{pairwise_contrast_matrix}
\title{Construct fused penalty structure matrix}
\usage{
pairwise_contrast_matrix(n.cols)
}
\arguments{
\item{n.cols}{\link{numeric}: Number of regression parameters, i.e. covariates}
}
\value{
K \link{matrix}: Fused penalty structure matrix of dimension (2^(P\emph{Q)) x (P}Q)
}
\description{
R-function constructing the fused penalty structure matrix of all pairwise differences for use in penalized regression
}
