% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_ll.R
\name{partial_ll}
\alias{partial_ll}
\title{partial_ll}
\usage{
partial_ll(X, d, beta, risksetlist)
}
\arguments{
\item{X}{\link{matrix}: Regression matrix of dimension n_obs x p_vars}

\item{d}{\link{data frame}: Data set with variables Tstart, Tstop, trans and status}

\item{beta}{\link{vector}: Regression parameter}

\item{Riskset}{\link{list}: Risk set list}
}
\value{
loglik \link{numeric}: Partial log-likelihood at beta
}
\description{
R-function calculating the partial log-likelihood function
}
