% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_matrix_K.R
\name{penalty_matrix_K}
\alias{penalty_matrix_K}
\title{penalty_matrix_K}
\usage{
penalty_matrix_K(P, Q, fused = "all", D = NULL, groups)
}
\arguments{
\item{P}{\link{numeric}: Number of regression parameters, i.e. covariates}

\item{Q}{\link{numeric}: Number of transitions}

\item{fused}{\link{character or matrix}: Character string/matrix indicating whether all pairwise differences ("all") or adjacent differences ("neighbors") or user-specific pairs (matrix) shall be penalized for the fused penalty}

\item{D}{\link{matrix}: Difference matrix}

\item{groups}{\link{vector}: Vector indicating group membership of regression parameters for the group penalty}
}
\value{
K \link{matrix}: General combined penalty matrix of dimension (P\emph{Q + s + P}Q) x (P*Q)
}
\description{
R-function constructing a combined penalty structure matrix for use in penalized regression
}
