% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_kappa.R
\name{S_kappa}
\alias{S_kappa}
\title{S_kappa}
\usage{
S_kappa(a, kappa)
}
\arguments{
\item{a}{\link{vector}: Numeric vector}

\item{kappa}{}
}
\value{
s \link{numeric}: Shrinked value
}
\description{
R-function implementing the vector soft thresholding operator S_kappa(a)
}
